osd: implement osd_distrust_data_digest option
This is a temporary workaround for a broken release (12.2.6) of luminous
that included a bug in WRIREFULL in which the data digest was not cleared
when bluestore was in use (and skip_data_digest == true).  The backport
2dd1265a54316da8980a377615ce2a19228c7aed was not included in 12.2.6.

A comprehensive solution would require some effort as we'd need to ignore
object digests until it is confirmed that all bad digests have been
scrubbed out of the system.  We may rely on operators doing this scrub
before disabling this option again.

In the meantime, this will fix the broken clusters.

Signed-off-by: Sage Weil <sage@redhat.com>