doc: use :command: for subcommands in ceph-bluestore-tool manpage
Older versions of Sphinx, such as the one in CentOS 7, do not render "..
option::" lines correctly if the option contains a hyphen but does not start
with a hyphen. And ceph-bluestore-tool appears to be the only Ceph manpage
affected by this bug.

Fixes: http://tracker.ceph.com/issues/24800
Signed-off-by: Nathan Cutler <ncutler@suse.com>
(cherry picked from commit 7c9229e7876dfc1bf6e415ed7cce924c3b50925d)