server: add RejectThreshold and AtLimitParam
add a tunable threshold for AtLimit::Reject such that requests are only
rejected once the limit is over this threshold. this is intended to help
balance the queue depth with the desire to enforce an upper bound on the
time spent waiting in the queue (and the memory use associated with
queued requests)

for example, setting this threshold to 1 second for a client with
limit=10 would allow that client to enqueue 10 requests before any are
rejected. with a threshold of 0, all but the first would be rejected and
require the client to resubmit

to simplify queue construction, a variant type AtLimitParam allows
construction with either an explicit AtLimit enum value, or with a
RejectThreshold value that implies AtLimit::Reject

Signed-off-by: Casey Bodley <cbodley@redhat.com>