mimic: qa/ceph-disk: fixes for OVH and add ceph-detect-init subsuite
With the move to OVH and the addition of RHEL as a supported distro, the
ceph-disk/basic suite was no longer green. This commit brings it back into
a green state.

Also, although ceph-disk/basic presumably exercises ceph-detect-init, it
seemed prudent to add a test case that runs it explicitly.

Can not be cherry-picked because ceph-disk and ceph-detect-init are no longer
present in master.

Reverts: 230c030fad091ae9c136a421e6206d3e8c9378f8
Fixes: https://tracker.ceph.com/issues/25031
Signed-off-by: Nathan Cutler <ncutler@suse.com>