rgw: civetweb frontend now has a dmc scheduler
The scheduler is maintained in the frontend itself right now instead of passing
it along from main. For this a function is added in rgw::dmclock::SyncScheduler
which will serve as the callback for SyncScheduler. This also holds a reference
to the perf counters so that the respective counters can be incremented on
granting a throttle unit. AtLimit behaviour is same as the AsyncVersion, where
AtLimit::Reject is followed. In rgw_process we currently reraise EAGAIN as
ERR_RATE_LIMITED.

Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>