mgr/influx: Use Queue to store points which need to be written
This allows us to multiplex data being send to Influx as we have
a configurable amount of workers sending data to Influx.

The main bottleneck for the performance seems to be fetching all
the perf counters using this code:

    self.get_all_perf_counters()

On a larger cluster, for example 2000 OSDs this can take about 20s
where flushing to Influx only takes 5s.

A 2000 OSD cluster generates about 100k data points on every run,
prior to using a Queue these would all be send to Influx in series
in that took over 15 seconds to complete.

Python Six is being used in the code to make sure it's compatible
with both Python 2 and 3.

Signed-off-by: Wido den Hollander <wido@42on.com>