mon: fix idempotent FS creation
In the case that our creation is a no-op, we should tolerate
objects in the metadata pool -- since we're not performing
any action, we can't do harm.

This case comes up if you have MDS daemons starting immediately
on filesystem creation: running the `fs new` command twice will fail
the second time because of the objects the MDS created.

Signed-off-by: John Spray <john.spray@redhat.com>