ceph-volume: run tests without waiting on ceph repos
This provides a new playbook that installs ceph on all nodes and
generates the ceph config. Then it will rsync ceph-volume from the
control node, or jenkins slave, to all testing vms before completing
the deployment. This means we can run tests on PRs without waiting
for repos to be built for the branch in the PR.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>
(cherry picked from commit 6d712ef959f7ea5385bc6bf9b9ade60662940369)