osd/OSD: fix project_pg_history vs osdmap injection race
The peering events used to be processed under the protection
of osd_lock, which as result can eliminate a race condition with
the new osdmap injection process.

However, the above case is not applicable any more as we now start
fast dispatching peering events.

Fix the potential race condition by passing in a sane, verified but
perhaps not necessarily the latest osdmap.
This should be fine because the peering process
will restart each time a new osdmap triggering a new pg-interval.

Fixes: http://tracker.ceph.com/issues/26970
Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>