mds/FSMap: specialize for WITH_SEASTAR
FSMap.{h,cc} is pulled in as a dependency of src/msg/Messages.cc, where
a handful of Messages types used by MDS and its client are referenced.
so we need to compile the referenced source files as well. another
option is to enable a subset of messages for different kinds of
server/client, but that needs more work.

Signed-off-by: Kefu Chai <kchai@redhat.com>