doc/dev/msgr2: remove stream concept, streamline auth
- Remove stream concept.  it was incompletely/poorly
specified, and we dno't actually want to implement it now. We
can do it later now that we have a banner that advertises
features more clearly.

- Simplify the auth messages a bit: make a MORE message that
isn't a request (or reply), and a final DONE.  DONE comes from
the server, not the client.  Drop the SET_METHOD which can be
rolled into the REQUEST message.

Signed-off-by: Sage Weil <sage@redhat.com>