vstart: set admin socket for RGW in conf
Currently ceph daemon cannot find rgw socket as client section sets the admin
socket to the tmp dir. whereas we actually write the socket in the out/
directory. Add this to ceph.conf, so that ceph daemon commands can find the
socket, avoiding the need for explicit path via --admin-daemon. When multiple
RGWs are created ceph daemon will only find the first configured RGW (as we are
only configuring one RGW per ceph.conf) and the others will have to be
explicitly passed via --admin-daemon or so.

Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>