bluestore/NVMe: use PCIe selector as the path name
as the latest SPDK deprecated and removed spdk_pci_get_device(),
we cannot use the PCI device's serial number as its identifier and the
filename of the file representing the device. in this change, we
are using the PCI device's selector instead, it is also used by SPDK as
part of the transport id.

Tested-by: Tone Zhang <tone.zhang@arm.com>
Signed-off-by: Kefu Chai <kchai@redhat.com>