client: fix fuse client can't write data due to don't have Fw capability
A file's wnat caps be pAsxXsxFsxcrwb if we open a file with O_RDWR flag
and the want caps are also recorded on MDS side.The MDS will grant the
caps to client according the want and the lock state machine.But if kill
session happen in mds due to the bad network the caps of all files will
be cleared. Whereafter the network back to normal the session will be
recreated but the wanted cap of all files will be lost.So, the client
can’t acquire the caps it needed and will waiting all the time.

Fixes:http://tracker.ceph.com/issues/36072

Signed-off-by:Guan yunfei <yunfei.guan@xtaotech.com>