osd/ECBackend: suppress 'Error -2 reading object' if EC fast reads
When fast reads are enabled, it's possible for the ordering of a shard
read to not be enforced with respect to writes that come after because
the read completes on the primary before all shards reply.  This can lead
to an ENOENT on the non-primary, and an ERR message in the cluster log,
even though everything is fine.  (The reply will go back to the primary
with the error but it will be ignored since the read has completed.)

Suppress the error message so we don't see these ERR messages in the
cluster log during the normal course of events.

Fixes: http://tracker.ceph.com/issues/26972
Signed-off-by: Sage Weil <sage@redhat.com>