mon: mon_osd_cache_size = 500 (from 10)
10 maps is too small to enable all mon sessions to keep abreast of the
latest maps, especially if OSDs are down for any period of time during an
upgrade.

Note that this is quite a bit larger, but the memory usage of the mon will
scale proportionally to the size of the cluster: 500 small osdmaps is not
a significant amount of RAM, while conversely having a large cache is
most important on a large cluster and those mons will generally have
plenty of RAM available.

Someday we should control this with a memory envelope like we do with the
OSDs, but that remains future work.

Signed-off-by: Sage Weil <sage@redhat.com>