rocksdb: increase default cache size to 512 MB
For filestore OSDs, this is probably a good idea anyway, and is generally
not going to be hugely impactful on the memory footprint (where users
have been told to provide 1 GB RAM per 1 TB storage for a long time now).

For bluestore OSDs, this value is meaningless as we're autotuning this
anyway.

For mons, this is a more reasonable default.

Signed-off-by: Sage Weil <sage@redhat.com>