common: remove unused 'avail' field from commands
This was introduced for the now-removed ceph-rest-api
gateway.  It enabled limiting certain commands
to be CLI-only or rest-only, but in practice almost
everything just said "cli,rest" here.

Now that ceph-rest-api is gone, let's remove this
field.

The CLI client code already tolerated the absence
of this field, so older CLI clients will not mind.

Signed-off-by: John Spray <john.spray@redhat.com>