rgw: admin command to cleanup stale bucket indexes from reshard
`radosgw-admin reshard cleanup --bucket={}`

This currently processes only a single bucket, it starts to reshard if the
marker differs from the entrypoint bucket id, indicating a resharded bucket.
Since a bucket may have been resharded multiple times, we loop through the next
bucket id and confirm that it is not the current and continue to purge bucket
indices until we reach the original bucket id.

Fixes: http://tracker.ceph.com/issues/24082
Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>