os: Transaction uses append_hole() to minimize bl:_buffers inflation.
This could be useful as the used `::append(const ptr&)` variant does
not provide any measures (like coalescing adjacent bptrs) to prevent
inflation of the internal structures of a bufferlist. That is, each
call was reflected in dynamic node allocation for the `_buffers` list.
Secondary effect of this behavior is increased iteration cost driven
by large number of elements likely scattered over memory.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>