ceph-kvstore-tool: rename repair -> destructive-repair
This is shown to corrupt otherwise healthy rocksdb databases.  Rename to
make it clear that it is generally not safe to run and shoud only be used
as a last resort.

Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit 8cc636951132c2ee93e32bfc999777e3da023dd7)

Conflicts:
	PendingReleaseNotes: drop this change as "repair" command did
not exist in luminous before this change.
	qa/workunits/cephtool/test_kvstore_tool.sh: drop this change,
as this test was not added before this change.
	src/tools/ceph_kvstore_tool.cc: trivial resolution.