Add a public controller with version endpoint
In order to display the version on the login page, we need to allow for
querying at least the version without authentication. I thought a decent
approach for this might be to create a controller at /api/public; if it
turns out that more endpoints need to be able to be used without
authentication, they could be put there.

I thought it might be useful to provide the number, hash and name
portions of the version in addition to the full raw string, but am not
attached to it.