server: should pass cost instead of 0 to RequestTag()
before this change, we always pass cost as delta to the constructor of
RequestTag, and use the default "1" for the constructed tag. after this
change, delta will be 0, and "cost" is passed to construct as "cost".

only delayed tag calculation is impacted by this issue. and we are
using immediate calc by default, and in Ceph's mClockQueue
implementation. so this issue is not critical.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>