rgw: reshard add: fail correctly on a non existant bucket
Currently the method that validates the reshard params does not fail correctly
in case of a non existant bucket causing reshard add to print an error message
but add the non existant bucket to the queue anyway. This will later be a cause
of EBUSY errors as this entry on the logshard  can never be processed.

Fixes: http://tracker.ceph.com/issues/36449
Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>
(cherry picked from commit c564b348a24edf2804a8f5f0d034f388be1c5c5f)
Signed-off-by: Jonathan Brielmaier <jbrielmaier@suse.de>